Moves delegates to individual files
