+++ piwigo/Settings/Upload/Auto-Upload/AutoUploadViewController.storyboard
